Refactor out BlockfaceLayer from SelectableBlockfaceLayer
